requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.0
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
ffmpeg-python==0.2.0
filelock==3.3.0
future==0.18.2
gunicorn==20.1.0
idna==3.2
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2==2.9.3
ptyprocess==0.7.0
py3-validate-email==1.0.2
Pygments==2.10.0
PyJWT==2.1.0
pytube==12.1.0
pytz==2021.3
requests==2.28.1
sqlparse==0.4.2
traitlets==5.1.1
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0


